Simple Address Book API
The three main classes for the address book design include:
-	Person
-	AddressBook
-	AddressBookManager
-	AddressBookConnection

For convenience, I am also adding two enumerations for the sort type and
print mode of the address list in a given address book:
public enum AddressListSortType { LastName, ZipCode };
public enum PrintMode { Database, Console }; // Disk or console output

Person:
It holds information about a person ÎíÎñ first name, last name, address,
city, state, zip, and phone number. The Person class does not expose
setting values only for first and last name properties.

AddressBook:
An AdressBook class object holds information about a particular instance
of an address book. The address book maintains/manages a list of Person
objects that constitute an address list. The main responsibilities of this
class include adding a new Person object to the address list, changing
ÎéÎímutableÎéÎí (all properties except first and last name) information on
a Person object, and deleting person object(s) from the maintained list.
It also sorts the maintained address list by last name, and by zip code.
And lastly, it saves the contents of the address list to a database, and
or prints to a console output.

AddressBookManager:
The AddressBookManager orchestrates/manages the core operations of the
application. It is responsible for creating, opening, saving and closing
address book(s) ÎíÎñ essentially, it manages the lifecycle of all address
books.

AddressBookConnection:
The AddressBookConnection class is an abstraction layer over consumable db
queries pertaining to retrieving specific information stored in the
address_book database (specifically in the address_info table). Exposed
functionality include retrieving, deleting and inserting entries into the
address_info table.

This also contains driver classes to test the functionality/behavior of
the above classes, and also JUnit tests to test the functionality of the
methods with the classes

[Includes JavaDocs]